feat(components): migrate image component to typescript
feature/typescript-image
