+++ src/components/Image/Image.docs.mdx
@@ -1,5 +1,5 @@
 import { Status, Props, Story } from '../../../.storybook/components';
-import Image from './Image';
+import { Image } from './Image';
 
 # Image
