+++ src/components/Image/Image.spec.tsx
@@ -15,6 +15,8 @@
 
 import React from 'react';
 
+import { create, renderToHtml, axe } from '../../util/test-utils';
+
 import Image from '.';
 
 describe('Image', () => {
