+++ src/components/Image/Image.story.tsx
@@ -13,6 +13,23 @@
  * limitations under the License.
  */
 
-import Image from './Image';
+import React from 'react';
 
-export default Image;
+import docs from './Image.docs.mdx';
+import { Image } from './Image';
+
+export default {
+  title: 'Components/Image',
+  component: Image,
+  parameters: {
+    docs: { page: docs },
+    jest: ['Image']
+  }
+};
+
+export const base = () => (
+  <Image
+    src="https://source.unsplash.com/random"
+    alt="A random image from Unsplash"
+  />
+);
